05_hackathon/congestion_tracker/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
httpx>=0.27.0
python-dotenv>=1.0.0
shiny>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.8.0


